habe die struktur im instruction menu aufgeräumt und für alle dialogfenster eigene components erstellt #47
