+++ ControlCenterApp/controlcenter-frontend/src/app/instruction-menu/dialogs/create-measure-dialog/create-measure-dialog.scss
@@ -1,0 +1,21 @@
+.full-width {
+  width: 100%;
+}
+
+.read-only-field .mat-input-element {
+  color: rgba(0, 0, 0, 0.6);
+}
+
+.preview-image {
+  max-width: 100%;
+  max-height: 220px;
+  margin-top: 12px;
+  border-radius: 4px;
+}
+
+.file-upload-row {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  margin-top: 12px;
+}
