+++ ControlCenterApp/controlcenter-frontend/src/app/instruction-menu/dialogs/create-plan-dialog/create-plan-dialog.scss
@@ -1,0 +1,7 @@
+.full-width {
+  width: 100%;
+}
+
+.read-only-field .mat-input-element {
+  color: rgba(0, 0, 0, 0.6);
+}
